move bind service to /extensions, skip digest for amp-bind-state
